Hopefully the effect bug is now done for good
It's not the moost elegant sulution, but it works... And i think that is
good enough right now becouse it was a pain in the but to first of all,
find the soursce of the poblem and then come up with a solution for it.
But as long as I haven't missed anything, again, it should work just
fine!